// src/app/modules/as/tablas/areas/zonas/service/zona-api-model-interfaces.ts
export interface IResponseZona {
	codZona: string;
	descripcion: string;
	activo: boolean;
	id: number;
	isDelete: boolean;
}

export interface IRequestCreateZona {
	codZona: string;
	descripcion: string;
}
export interface IHomeZona {
	id: number;
	codZona: string;
	description: string;
	activo: boolean;
}